all pages are now protected

+++ update/delete.php
@@ -5,36 +5,38 @@ include_once('../system/config.php');
 
 include_once('../templates/content.php');
 
-getHeader("Sqits form-delete", "Update verwijderen");
+if (isset($_SESSION['id'])) {
 
-checkRole("admin");
+    getHeader("Sqits form-delete", "Update verwijderen");
 
-//	$id = $_GET['id'];
-//	$_SESSION['id']['user_id'];
-    
-	if(@$_GET['action'] == "delete")
-	{	
-        
-        try
-        {
-            $query=$conn->prepare("
+    checkRole("admin");
+
+    if (@$_GET['action'] == "delete") {
+
+        try {
+            $query = $conn->prepare("
                         DELETE FROM `update` WHERE `update_id` = :id");
             $query->execute(array(
                 'id' => $_GET['id']
             ));
             header('Location: index.php');
-        }
-        catch(PDOException $e)
-        {
+        } catch (PDOException $e) {
             $sMsg = '<p>
-                    Regelnummer: '.$e->getLine().'<br />
-                    Bestand: '.$e->getFile().'<br />
-                    Foutmelding: '.$e->getMessage().'
+                    Regelnummer: ' . $e->getLine() . '<br />
+                    Bestand: ' . $e->getFile() . '<br />
+                    Foutmelding: ' . $e->getMessage() . '
                 </p>';
 
             trigger_error($sMsg);
-        }           
+        }
     }
-    
-	getFooter();	
+
+    getFooter();
+
+}else {
+    echo "please login first on login page";
+    header("Location:../login.php");
+    exit;
+}
+
 ?>
